Fixes #152 Changed the regular expression used for removing unused parts of the interval string to not drop zero digits at the end of integer values
